+++ src/ui/editor.ts
@@ -6,7 +6,6 @@ import { Events } from '../events';
 import { AboutPopup } from './about-popup';
 import { BottomToolbar } from './bottom-toolbar';
 import { CameraInfoOverlay } from './camera-info-overlay';
-import { ColorPanel } from './color-panel';
 import { ExportPopup } from './export-popup';
 import { ImageSettingsDialog } from './image-settings-dialog';
 import { i18n } from './localization';
@@ -27,6 +26,7 @@ import { TimelinePanel } from './timeline-panel';
 import { Tooltips } from './tooltips';
 import { VideoSettingsDialog } from './video-settings-dialog';
 import { ViewCube } from './view-cube';
+import { ViewOptionsPanel } from './view-options-panel';
 import { version } from '../../package.json';
 
 // ts compiler and vscode find this type, but eslint does not
@@ -95,7 +95,7 @@ class EditorUI {
         // bottom toolbar
         const scenePanel = new ScenePanel(events, tooltips);
         const settingsPanel = new SettingsPanel(events, tooltips);
-        const colorPanel = new ColorPanel(events, tooltips);
+        const viewOptionsPanel = new ViewOptionsPanel(events, tooltips);
         const bottomToolbar = new BottomToolbar(events, tooltips);
         const rightToolbar = new RightToolbar(events, tooltips);
         const modeToggle = new ModeToggle(events, tooltips);
@@ -109,12 +109,15 @@ class EditorUI {
         canvasContainer.append(perfOverlay);
         canvasContainer.append(toolsContainer);
         canvasContainer.append(scenePanel);
-        canvasContainer.append(settingsPanel);
-        canvasContainer.append(colorPanel);
         canvasContainer.append(bottomToolbar);
         canvasContainer.append(rightToolbar);
         canvasContainer.append(modeToggle);
         canvasContainer.append(menu);
+
+        // the option popups come after the toolbars so their select dropdowns,
+        // which can extend past the panel bounds, paint above them
+        canvasContainer.append(settingsPanel);
+        canvasContainer.append(viewOptionsPanel);
 
         // view axes container
         const viewCube = new ViewCube(events);
