+++ src/ui/right-toolbar.ts
@@ -6,11 +6,11 @@ import { i18n } from './localization';
 import cameraFrameSelectionSvg from './svg/camera-frame-selection.svg';
 import cameraResetSvg from './svg/camera-reset.svg';
 import centersSvg from './svg/centers.svg';
-import colorPanelSvg from './svg/color-panel.svg';
 import flyCameraSvg from './svg/fly-camera.svg';
 import orbitCameraSvg from './svg/orbit-camera.svg';
 import ringsSvg from './svg/rings.svg';
 import showHideSplatsSvg from './svg/show-hide-splats.svg';
+import viewOptionsSvg from './svg/view-options.svg';
 import { Tooltips } from './tooltips';
 
 const createSvg = (svgString: string) => {
@@ -61,8 +61,8 @@ class RightToolbar extends Container {
             class: 'right-toolbar-button'
         });
 
-        const colorPanel = new Button({
-            id: 'right-toolbar-color-panel',
+        const viewOptions = new Button({
+            id: 'right-toolbar-view-options',
             class: 'right-toolbar-toggle'
         });
 
@@ -83,7 +83,7 @@ class RightToolbar extends Container {
         flyMode.dom.appendChild(createSvg(flyCameraSvg));
         cameraFrameSelection.dom.appendChild(createSvg(cameraFrameSelectionSvg));
         cameraReset.dom.appendChild(createSvg(cameraResetSvg));
-        colorPanel.dom.appendChild(createSvg(colorPanelSvg));
+        viewOptions.dom.appendChild(createSvg(viewOptionsSvg));
 
         this.append(ringsModeToggle);
         this.append(showHideSplats);
@@ -94,7 +94,7 @@ class RightToolbar extends Container {
         this.append(cameraFrameSelection);
         this.append(cameraReset);
         this.append(new Element({ class: 'right-toolbar-separator' }));
-        this.append(colorPanel);
+        this.append(viewOptions);
         this.append(options);
 
         // Helper to compose localized tooltip text with shortcut
@@ -116,7 +116,7 @@ class RightToolbar extends Container {
         tooltips.register(flyMode, tooltip('tooltip.right-toolbar.fly-camera', 'camera.toggleControlMode'), 'left');
         tooltips.register(cameraFrameSelection, tooltip('tooltip.right-toolbar.frame-selection', 'camera.focus'), 'left');
         tooltips.register(cameraReset, tooltip('tooltip.right-toolbar.reset-camera', 'camera.reset'), 'left');
-        tooltips.register(colorPanel, tooltip('tooltip.right-toolbar.colors'), 'left');
+        tooltips.register(viewOptions, tooltip('tooltip.right-toolbar.view-options'), 'left');
         tooltips.register(options, tooltip('tooltip.right-toolbar.settings'), 'left');
 
         // add event handlers
@@ -130,7 +130,7 @@ class RightToolbar extends Container {
         flyMode.on('click', () => events.fire('camera.setControlMode', 'fly'));
         cameraFrameSelection.on('click', () => events.fire('camera.focus'));
         cameraReset.on('click', () => events.fire('camera.reset'));
-        colorPanel.on('click', () => events.fire('colorPanel.toggleVisible'));
+        viewOptions.on('click', () => events.fire('viewPanel.toggleVisible'));
         options.on('click', () => events.fire('settingsPanel.toggleVisible'));
 
         events.on('camera.mode', (mode: string) => {
@@ -148,8 +148,8 @@ class RightToolbar extends Container {
             flyMode.class[mode === 'fly' ? 'add' : 'remove']('active');
         });
 
-        events.on('colorPanel.visible', (visible: boolean) => {
-            colorPanel.class[visible ? 'add' : 'remove']('active');
+        events.on('viewPanel.visible', (visible: boolean) => {
+            viewOptions.class[visible ? 'add' : 'remove']('active');
         });
 
         events.on('settingsPanel.visible', (visible: boolean) => {
