+++ src/ui/scene-panel.ts
@@ -1,6 +1,7 @@
-import { Container, Element, Label } from '@playcanvas/pcui';
+import { Container, Label } from '@playcanvas/pcui';
 
 import { Events } from '../events';
+import { ColorPanel } from './color-panel';
 import { i18n } from './localization';
 import { SplatList } from './splat-list';
 import sceneImportSvg from './svg/import.svg';
@@ -116,10 +117,7 @@ class ScenePanel extends Container {
         this.append(splatListContainer);
         this.append(transformHeader);
         this.append(new Transform(events));
-        this.append(new Element({
-            class: 'panel-header',
-            height: 20
-        }));
+        this.append(new ColorPanel(events, tooltips));
     }
 }
