+++ src/ui/settings-panel.ts
@@ -2,17 +2,18 @@ import { BooleanInput, Button, ColorPicker, Container, Label, SelectInput, Slide
 import { Color } from 'playcanvas';
 
 import { Events } from '../events';
-import type { GridPlane } from '../infinite-grid';
-import { ShortcutManager } from '../shortcut-manager';
 import { i18n } from './localization';
+import resetSvg from './svg/edit-undo.svg';
 import { Tooltips } from './tooltips';
 
+// application preferences: set-and-forget options, as opposed to the viewport
+// state that lives in the view options panel
 class SettingsPanel extends Container {
     constructor(events: Events, tooltips: Tooltips, args = {}) {
         args = {
             ...args,
             id: 'settings-panel',
-            class: 'panel',
+            class: ['panel', 'options-panel'],
             hidden: true
         };
 
@@ -30,7 +31,7 @@ class SettingsPanel extends Container {
         });
 
         const icon = new Label({
-            text: '\uE403',
+            text: '\uE283',
             class: 'panel-header-icon'
         });
 
@@ -41,6 +42,30 @@ class SettingsPanel extends Container {
 
         header.append(icon);
         header.append(label);
+
+        // section bars share the panel-header styling, like the scene
+        // manager's transform header
+        const sectionHeader = (key: string) => {
+            const section = new Container({
+                class: ['panel-header', 'options-panel-section']
+            });
+            const sectionLabel = new Label({
+                class: 'panel-header-label'
+            });
+            i18n.bindText(sectionLabel, key);
+            section.append(sectionLabel);
+            return section;
+        };
+
+        // toggle rows flip on a click anywhere in the row, not just the switch
+        const rowToggles = (row: Container, toggle: BooleanInput) => {
+            row.class.add('options-panel-row-clickable');
+            row.dom.addEventListener('click', (event: MouseEvent) => {
+                if (toggle.enabled && !toggle.dom.contains(event.target as Node)) {
+                    toggle.value = !toggle.value;
+                }
+            });
+        };
 
         // language
 
@@ -253,67 +278,6 @@ class SettingsPanel extends Container {
         cameraFlySpeedRow.append(cameraFlySpeedLabel);
         cameraFlySpeedRow.append(cameraFlySpeedSlider);
 
-        // centers size
-
-        const centersSizeRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const centersSizeLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(centersSizeLabel, 'panel.settings.center-size');
-
-        const centersSizeSlider = new SliderInput({
-            class: 'settings-panel-row-slider',
-            min: 0,
-            max: 10,
-            precision: 1,
-            value: 2
-        });
-
-        centersSizeRow.append(centersSizeLabel);
-        centersSizeRow.append(centersSizeSlider);
-
-        // centers gaussian color
-        const centersColorRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const centersColorLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(centersColorLabel, 'panel.settings.use-splat-colors');
-
-        const centersColorToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        centersColorRow.append(centersColorLabel);
-        centersColorRow.append(centersColorToggle);
-
-        // outline selection
-
-        const outlineSelectionRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const outlineSelectionLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(outlineSelectionLabel, 'panel.settings.outline-selection');
-
-        const outlineSelectionToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        outlineSelectionRow.append(outlineSelectionLabel);
-        outlineSelectionRow.append(outlineSelectionToggle);
-
         // stochastic alpha
 
         const stochasticRow = new Container({
@@ -339,182 +303,36 @@ class SettingsPanel extends Container {
         stochasticRow.append(stochasticLabel);
         stochasticRow.append(stochasticSelection);
 
-        // frame timings overlay
-
-        const perfOverlayRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const perfOverlayLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(perfOverlayLabel, 'panel.settings.perf-overlay');
-
-        const perfOverlayToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        perfOverlayRow.append(perfOverlayLabel);
-        perfOverlayRow.append(perfOverlayToggle);
-
-        // show grid
-
-        const showGridRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const showGridLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(showGridLabel, 'panel.settings.show-grid');
-
-        const showGridToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: true
-        });
-
-        showGridRow.append(showGridLabel);
-        showGridRow.append(showGridToggle);
-
-        // grid plane
-
-        const gridPlaneRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const gridPlaneLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(gridPlaneLabel, 'panel.settings.grid-plane');
-
-        const gridPlaneSelection = new SelectInput({
-            class: 'settings-panel-row-select',
-            defaultValue: 'xz',
-            options: [
-                { v: 'xz', t: 'XZ' },
-                { v: 'xy', t: 'XY' },
-                { v: 'yz', t: 'YZ' }
-            ]
-        });
-
-        gridPlaneRow.append(gridPlaneLabel);
-        gridPlaneRow.append(gridPlaneSelection);
-
-        // show bound
-
-        const showBoundRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const showBoundLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(showBoundLabel, 'panel.settings.show-bounding-box');
-
-        const showBoundToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: true
-        });
-
-        showBoundRow.append(showBoundLabel);
-        showBoundRow.append(showBoundToggle);
-
-        // show dimensions
-
-        const showBoundDimensionsRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const showBoundDimensionsLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(showBoundDimensionsLabel, 'panel.settings.show-dimensions');
-
-        const showBoundDimensionsToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        showBoundDimensionsRow.append(showBoundDimensionsLabel);
-        showBoundDimensionsRow.append(showBoundDimensionsToggle);
-
-        // show camera poses
-
-        const showCameraPosesRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const showCameraPosesLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(showCameraPosesLabel, 'panel.settings.show-camera-poses');
-
-        const showCameraPosesToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        showCameraPosesRow.append(showCameraPosesLabel);
-        showCameraPosesRow.append(showCameraPosesToggle);
-
-        // show camera info
-
-        const showCameraInfoRow = new Container({
-            class: 'settings-panel-row'
-        });
-
-        const showCameraInfoLabel = new Label({
-            class: 'settings-panel-row-label'
-        });
-        i18n.bindText(showCameraInfoLabel, 'panel.settings.show-camera-info');
-
-        const showCameraInfoToggle = new BooleanInput({
-            type: 'toggle',
-            class: 'settings-panel-row-toggle',
-            value: false
-        });
-
-        showCameraInfoRow.append(showCameraInfoLabel);
-        showCameraInfoRow.append(showCameraInfoToggle);
-
         // reset preferences to defaults
 
         const resetRow = new Container({
-            class: 'settings-panel-row'
+            class: ['settings-panel-row', 'options-panel-row-footer']
         });
 
         const resetButton = new Button({
             class: 'settings-panel-row-button'
         });
         i18n.bindText(resetButton, 'panel.settings.reset');
+        // the icon renders as a css mask (a child svg would be wiped by the
+        // button's text setter whenever the language changes)
+        resetButton.dom.style.setProperty('--icon', `url("${resetSvg}")`);
 
         resetRow.append(resetButton);
 
+        rowToggles(fovDollyRow, fovDollyToggle);
+
         this.append(header);
+        this.append(sectionHeader('panel.settings.section-application'));
         this.append(languageRow);
         this.append(clrRow);
+        this.append(sectionHeader('panel.settings.section-rendering'));
+        this.append(stochasticRow);
         this.append(tonemappingRow);
+        this.append(shBandsRow);
+        this.append(sectionHeader('panel.settings.section-camera'));
+        this.append(cameraFlySpeedRow);
         this.append(fovRow);
         this.append(fovDollyRow);
-        this.append(shBandsRow);
-        this.append(cameraFlySpeedRow);
-        this.append(centersSizeRow);
-        this.append(centersColorRow);
-        this.append(outlineSelectionRow);
-        this.append(showGridRow);
-        this.append(gridPlaneRow);
-        this.append(showBoundRow);
-        this.append(showBoundDimensionsRow);
-        this.append(showCameraPosesRow);
-        this.append(showCameraInfoRow);
-        this.append(stochasticRow);
-        this.append(perfOverlayRow);
         this.append(resetRow);
 
         // handle panel visibility
@@ -538,7 +356,7 @@ class SettingsPanel extends Container {
             setVisible(this.hidden);
         });
 
-        events.on('colorPanel.visible', (visible: boolean) => {
+        events.on('viewPanel.visible', (visible: boolean) => {
             if (visible) {
                 setVisible(false);
             }
@@ -552,27 +370,6 @@ class SettingsPanel extends Container {
 
         shBandsSlider.on('change', (value: number) => {
             events.fire('view.setBands', value);
-        });
-
-        // splat size
-
-        events.on('camera.splatSize', (value: number) => {
-            centersSizeSlider.value = value;
-        });
-
-        centersSizeSlider.on('change', (value: number) => {
-            events.fire('camera.setSplatSize', value);
-            events.fire('camera.setOverlay', true);
-            events.fire('camera.setMode', 'centers');
-        });
-
-        // centers gaussian color
-        events.on('view.centersUseGaussianColor', (value: boolean) => {
-            centersColorToggle.value = value;
-        });
-
-        centersColorToggle.on('change', (value: boolean) => {
-            events.fire('view.setCentersUseGaussianColor', value);
         });
 
         // camera speed
@@ -595,16 +392,6 @@ class SettingsPanel extends Container {
             events.fire('camera.setFovDolly', value);
         });
 
-        // outline selection
-
-        events.on('view.outlineSelection', (value: boolean) => {
-            outlineSelectionToggle.value = value;
-        });
-
-        outlineSelectionToggle.on('change', (value: boolean) => {
-            events.fire('view.setOutlineSelection', value);
-        });
-
         // stochastic alpha
 
         events.on('view.stochastic', (value: string) => {
@@ -613,76 +400,6 @@ class SettingsPanel extends Container {
 
         stochasticSelection.on('change', (value: string) => {
             events.fire('view.setStochastic', value);
-        });
-
-        // frame timings overlay
-
-        events.on('view.perfOverlay', (value: boolean) => {
-            perfOverlayToggle.value = value;
-        });
-
-        perfOverlayToggle.on('change', (value: boolean) => {
-            events.fire('view.setPerfOverlay', value);
-        });
-
-        // show grid
-
-        events.on('grid.visible', (visible: boolean) => {
-            showGridToggle.value = visible;
-        });
-
-        showGridToggle.on('change', () => {
-            events.fire('grid.setVisible', showGridToggle.value);
-        });
-
-        // grid plane
-
-        events.on('grid.plane', (plane: GridPlane) => {
-            gridPlaneSelection.value = plane;
-        });
-
-        gridPlaneSelection.on('change', (value: GridPlane) => {
-            events.fire('grid.setPlane', value);
-        });
-
-        // show bound
-
-        events.on('camera.bound', (visible: boolean) => {
-            showBoundToggle.value = visible;
-        });
-
-        showBoundToggle.on('change', () => {
-            events.fire('camera.setBound', showBoundToggle.value);
-        });
-
-        // show dimensions
-
-        events.on('camera.boundDimensions', (visible: boolean) => {
-            showBoundDimensionsToggle.value = visible;
-        });
-
-        showBoundDimensionsToggle.on('change', () => {
-            events.fire('camera.setBoundDimensions', showBoundDimensionsToggle.value);
-        });
-
-        // show camera poses
-
-        events.on('camera.showPoses', (visible: boolean) => {
-            showCameraPosesToggle.value = visible;
-        });
-
-        showCameraPosesToggle.on('change', () => {
-            events.fire('camera.setShowPoses', showCameraPosesToggle.value);
-        });
-
-        // show camera info
-
-        events.on('camera.showInfo', (visible: boolean) => {
-            showCameraInfoToggle.value = visible;
-        });
-
-        showCameraInfoToggle.on('change', () => {
-            events.fire('camera.setShowInfo', showCameraInfoToggle.value);
         });
 
         // background color
@@ -736,11 +453,6 @@ class SettingsPanel extends Container {
         });
 
         // tooltips
-        const shortcutManager: ShortcutManager = events.invoke('shortcutManager');
-        const shortcut = shortcutManager.formatShortcut('grid.toggleVisible');
-        tooltips.register(showGridLabel, () => i18n.formatTooltipWithShortcut(i18n.t('panel.settings.show-grid'), shortcut), 'left');
-        const cameraInfoShortcut = shortcutManager.formatShortcut('camera.toggleShowInfo');
-        tooltips.register(showCameraInfoLabel, () => i18n.formatTooltipWithShortcut(i18n.t('panel.settings.show-camera-info'), cameraInfoShortcut), 'left');
         tooltips.register(bgClrPicker, () => i18n.t('panel.settings.background-color'), 'left');
         tooltips.register(selectedClrPicker, () => i18n.t('panel.settings.selected-color'), 'top');
         tooltips.register(unselectedClrPicker, () => i18n.t('panel.settings.unselected-color'), 'top');
